pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel-monitor"
version = "0.1.0"
description = "Production attribution and contract tightening — watches logs, attributes errors via PACT keys, spawns fixer agents."
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40"]
cloudwatch = ["boto3>=1.34"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
]

[project.scripts]
sentinel = "sentinel.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
